Added ToSchema instances
Unfortunately, I don't think the ToSchema class in the module Schema
supports having Schema's for sum types which take arguments (the type
definition appears to be not flexible enough). In this case we can factor
out the the sum type part since both sides of the Either hold the same
data.