Include sidechainParams in signatures to prevent cross-chain replay attacks
It is a common practice to include parameters that uniquely identify given chain in the signatures, in our case a particular sidechain is identified by `SidechainParams` data structure. By doing that we also prevent any possible cross-chain replay attacks as the signatures are now bound to a particular chain.

Apart from that I also changed the serialization schema for signatures to cbor, since that should be easier to calculate for all of us.

Last but not least, I changed the type of sidechainPubKey references, since sometimes it was incorrectly set to `PubKey` which is an EDDSA key.