Implement versioning system and FUEL proxy mechanism (#8)
* Implement versioning system and fuel proxy minting policy

This implements SIP-01.

We introduce a script versioning system (TrustlessSidechain.Versioning modules
in onchain and offchain code) that acts both as an oracle of available scripts
as well a cache of reference scripts. Scripts are initially versioned during
sidechain initialization. These scripts can later be fetched in the on-chain
code. Importantly, these allows us to do away with passing currency symbols /
minting policies as script parameters. For example, FUELMint data type is no
longer necessary since we can fetch desired merkleRootTokenCurrencySymbol and
dsKeyCurrencySymbol from the versioning system. (An important exception to this
is the parameterization of versioning validator with versioning token currency
symbol.) Each script has its unique identifier. In the on-chain code these
identifiers are integers so that it is possible to add new script IDs without
any changes to existing code, and thus script addresses. However, in the
off-chain code we use custom ToData/FromData instances to decode these integers
into an ADT. It is a lot more convenient to use ADT constructors than magic
numbers, but care must be taken to keep the identifiers aligned between on-chain
and off-chain code. (In fact this would be true even without an ADT.)

With the versioning system in place we introduce a FUEL proxy mechanism, which
allows to modify fuel minting and burning policies without changing fuel
currency symbol. Importantly, this mechanism requires two separate policies -
one for minting and one for burning - to be stored in the versioning
system. This means that the previously existing FUELMintingPolicy has been
separated into two scripts now stored in TrustlessSidechain.FUELMintingPolicy.V1
and TrustlessSidechain.FUELBurningPolicy.V1. Importantly, and somewhat
counter-intuitively, burning policy also mints tokens. More importantly, the
recipient address is now placed in FUELProxyPolicy redeemer, not the redeemer of
the underlying policy. This probably requires changes in the bridge.

See Test.FUELProxyPolicy for integration tests that demonstrate fuel proxy
mechanism in action. For demonstration purposes we created V2 scripts for
minting and burning. These allow to mint tokens unconditionally. They serve
demonstration purposes only and shouldn't really be in the library, but we had
to add them to demonstrate that the whole thing works. In the on-chain code this
dummy policy script is placed in the test code inside
Test.TrustlessSidechain.DummyMintingPolicy module, but in the off-chain code
this has to be an integral part of the code.

Furthermore, this PR adds a simple governance mechanism, implemented as a single
master-key whose signature is required to perform certain sidechain changes. In
particular, governance authority signature is required to interact with the
versioning system (insertion/invalidation/update). Governance authority is
specified using short -g or long --governance-authority command line option or
in a config file via governanceAuthority entry. Expected value is a public key
hash.

Lastly, there is a command line interface for the versioning system, which
should enable end-to-end testing. In particular, cli allows to:

  * Claim and burn tokens using specified script version. Use claim-v1, burn-v1,
    claim-v2, burn-v2 commands instead of previous claim and burn.

  * Insert new version of scripts using insert-version command.

  * Invalidate existing script version invalidate-version command.

  * Update from one version to another update-version command.

Note that managing script versions requires signature from the governance
authority. At the moment governance authority must coincide with the private key
used to sign sidechain transactions. Also, at the moment only versions 1 and 2
are supported, where version 2 uses dummy demonstration policies mentioned
earlier.

Last but not least, this commit resolves all but one compilation warning.  Fixes #46