Specialize script utility functions
The two changed utility functions required that we pass a homogenous array of
serializable parameters.  This might have been fine in few specific places these
functions were used at*, but in general parameters to the script are not of the
same type and so it is impossible to pass them in a homogenous array.  We
therefore change these utility functions such that they take an array of already
serialized data.  This will allow to use these two utility functions throughout
the code.

*) In fact in the places where these functions are used the data is already
   passed as serialized.