Remove duplicate code
Apparently, there were two nearly identical function for balancing, signing, and
submitting a transaction.  We remove submitAndAwaitTx, because it had slightly
worse error handling.  We also modify the remaining balanceSignAndSubmit to take
lookups and constraints as a record, since this allows composing with functions
that build a record of lookups and constraints.