Introduce a map between script IDs and serialized scripts
Previously, when deserializing scripts from a text envelope we would directly
pass a string containing the script encoded as CBOR.  This works, but in case of
errors we didn't have access to information which script was being decoded.
This patch introduces a map from script IDs to serialized raw scripts.  Now,
when deserializing a script we pass its ID, which enables further improvements
in reporting errors.  Note that this change applies only to scripts used inside
the library.  For PoC scripts we use the old method of directly passing the CBOR
since we don't care about accurate bug reporting in tests.

Corollary: every serialized script requires and ID, even if it is not versioned.