Flatten error representation in off-chain code
Currently existing hierarchy of errors, i.e. the split between input and
internal errors, made no sense.  For many errors it is hard to say to which
category they should belong.  Many fall into two.  For example, when signatures
mismatch is it because something went wrong internally in the sidechain
representation, or did the user simply made a typo on the command line?  We
therefore give up on this artificial hierarchy in favour of putting all errors
into the same data type.