Add dedicated data type to store minting policy and curency symbol
It is a common pattern throughout the code to simultaneously handle currency
symbols and minting policies.  Until now each module used an anonymous record to
store a MintingPolicy and a CurrencySymbol.  An up-side of that approach was
that the record field names were informative, i.e.

  { checkpointCurrencySymbol, checkpointMintingPolicy }

is a bit more informative than a more generic:

  { currencySymbol, mintingPolicy }

However, this approach results in code repetition, missed abstraction
opportunities, and unnecessary boilerplate (c.f. #648).  We therefore introduce
a dedicated data type:

  type CurrencyInfo =
    { mintingPolicy ∷ MintingPolicy
    , currencySymbol ∷ CurrencySymbol
    }

that we use whenever calculating minting policy and currency symbol of a
deserialized script.  So previously we would write:

  { checkpointCurrencySymbol, checkpointMintingPolicy } <- getCheckpointPolicy

and now we write:

  { currencySymbol, mintingPolicy } <- getCheckpointPolicy

Note that the informativeness mentioned before can be recovered by writing:

  checkpointCurrencyInfo <- getCheckpointPolicy

and then using `checkpointCurrencyInfo.currencySymbol` and
`checkpointCurrencyInfo.mintingPolicy`.