Measure the size of Plutus scripts in terms of Flat encoding
We previously measured size of our scripts in terms of AST node count.  This is
wrong, because it only aproximates the actual size.  For example, a string
constant is always of the size of 1 in terms of AST nodes, but its actual size
in bytes depends on the constant's length.  This change switches to measuring
size of flat-encoded scripts.  Flat encoding is what ultimately is submitted to
Cardano network, so this seems like a good step forward.