Fix unorderedEq
The existing implementation of unorderedEq crucially relied on the fact that
both keys and values are returned in the same order.  This happens to be the
case for the currently existing implementation based on lists.  However, this is
not an invariant we should rely on.  Instead, let's do something simpler and
just unwrap the underlying lists and compare them after sorting.  This is
simpler, faster, and eliminates the extra assumption about alignment of keys and
values functions.  The downside is that we unwrap the abstraction and rely
directly on the implementation.