Remove redundant code
Initialization routine for the candidate permission token mechanism had code for
versioning the required minting policies and validtors, except that it doesn't
really require any.  I decided to remove this "speculative" code to have less
code to maintain.  If we ever need to version any scripts for candidate
permission tokens, we will then restore this.