Weaken the governance check
Don't demand that exactly one governance token must be minted in order to
consider a transaction approved.  There is really no need for such a restriction
and it is conceivable that future versions of governance might want to mint more
than one token to communicate extra information for external observers.