Provide nix-shell environment for offchain part
This adds a whole bunch of files providing the dependencies needed
for building the offchain component. This includes the following
external dependencies/overlays

* purescript-overlay
* cardano-node
* kupo
* ogmios

NOTE: Currently there are no upstream nix expressions for kupo and
ogmios. We fetch zips from the the github release for both of them.

NOTE: This commit only provides zip files for x86-64-linux, macos
will be added in a follow-up commit