provide check-raw-scripts.sh to validate RawScipts
The check-raw-scripts.sh script invokes make to build the output files
to a temporary directory and then compares them to the one that have
been committed to the repository.

The Makefile has been changeds slightly for this in oder to allow
overriding of the output location of the raw scripts.